cooldown interaction changed to be relationship between giver and receiver

+++ cogs/karma/producer.py
@@ -23,7 +23,7 @@ class KarmaProducer(commands.Cog):
         self.bot = bot
         self.karma_service = karma_service
         self.blocker_service = blocker_service
-        self._members_on_cooldown = defaultdict(list)
+        self._members_on_cooldown = defaultdict(lambda: defaultdict(list))
 
     # give karma if message has thanks and correct mentions
     @guild_only()
@@ -33,17 +33,7 @@ class KarmaProducer(commands.Cog):
         if not message.author.bot:
             if await self.validate_message(message):
                 if self.blocker_service.find_member(Member(str(guild_id), message.author.id)) is None:
-                    if message.author.id not in self._members_on_cooldown[guild_id]:
-                        await self.give_karma(message, message.guild, True)
-                    else:
-                        log.info('Sending configured cooldown response to {} in guild {}'
-                                 .format(message.author.id, guild_id))
-                        if str(config['karma']['time-emote']).lower() == "true":
-                            await message.add_reaction('🕒')
-                        if str(config['karma']['time-message']).lower() == "true":
-                            await self.bot.get_channel(message.channel.id) \
-                                .send('Sorry {}, your karma needs time to recharge'
-                                      .format(message.author.mention))
+                    await self.give_karma(message, message.guild, True)
                 else:
                     if str(config['blacklist']['message']).lower() == 'true':
                         log.info('Sending Blacklist dm to {} in guild {}'.format(message.author.id, guild_id))
@@ -97,23 +87,30 @@ class KarmaProducer(commands.Cog):
     # logged to a configured channel with member name & discriminator, optionally with nickname
     # cooldown author after successfully giving karma
     async def give_karma(self, message: discord.Message, guild: discord.Guild, inc: bool):
-        karma_given = 0
         for mention in set(message.mentions):
             member = mention
             if member.id != message.author.id and member.id != self.bot.user.id and not \
                     self.bot.get_user(member.id).bot:
-                log.info('{} gave karma to {} in guild {} with inc {}'
-                         .format(message.author.id, member.id, guild.id, inc))
-                karma_member = KarmaMember(guild.id, member.id, message.channel.id, message.id)
-                if inc:
-                    self.karma_service.upsert_karma_member(karma_member)
-                    karma_given += 1
-                    await self.notify_member(message, member)
+                if member.id not in self._members_on_cooldown[guild.id][message.author.id]:
+                    karma_member = KarmaMember(guild.id, member.id, message.channel.id, message.id)
+                    if inc:
+                        self.karma_service.upsert_karma_member(karma_member)
+                        await self.cooldown_user(guild.id, message.author.id, member.id)
+                        await self.notify_member(message, member)
+                    else:
+                        karma_member.karma = 1
+                        self.karma_service.delete_karma_member(karma_member)
+                        log.info('{} gave karma to {} in guild {} with inc {}'
+                             .format(message.author.id, member.id, guild.id, inc))
                 else:
-                    karma_member.karma = 1
-                    self.karma_service.delete_karma_member(karma_member)
-        if karma_given > 0:
-            await self.cooldown_user(guild.id, message.author.id)
+                    log.info('Sending configured cooldown response to {} in guild {}'
+                             .format(message.author.id, guild.id))
+                    if str(config['karma']['time-emote']).lower() == "true":
+                        await message.add_reaction('🕒')
+                    if str(config['karma']['time-message']).lower() == "true":
+                        await self.bot.get_channel(message.channel.id) \
+                            .send('Sorry {}, your karma for {} needs time to recharge'
+                                  .format(message.author.mention, member.name))
 
     # notify user about successful karma gain
     async def notify_member(self, message, member):
@@ -139,11 +136,11 @@ class KarmaProducer(commands.Cog):
             await message.add_reaction('👍')
 
     # create new timer and add the user to it
-    async def cooldown_user(self, guild_id: int, member_id: int) -> None:
-        self._members_on_cooldown[guild_id].append(member_id)
+    async def cooldown_user(self, guild_id: int, giver_id: int, receiver_id: int) -> None:
+        self._members_on_cooldown[guild_id][giver_id].append(receiver_id)
         await KarmaSingleActionTimer(self.remove_from_cooldown, int(config['cooldown']),
-                                     guild_id, member_id).start()
+                                     guild_id, giver_id, receiver_id).start()
 
     # remove user from cooldown after time runs out
-    async def remove_from_cooldown(self, guild_id: int, member_id: int) -> None:
-        self._members_on_cooldown[guild_id].remove(member_id)
+    async def remove_from_cooldown(self, guild_id: int, giver_id: int, receiver_id: int) -> None:
+        self._members_on_cooldown[guild_id][giver_id].remove(receiver_id)
